Split shuffle.py into py2s, py1s algo files.
